+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
  - recoded using classes  
  - change of savedata format  
  - moved from CLI to Kivy app  
+ - scheduling is now based on Berger Tables instead of an iterative discovery algorithm  
 
 ## [0.1.0]
 #### Add